Resolve relative JSON references
In order to resolve relative JSON references we need to pass a proper
base URI to RefResolver. By design, JSON schema defines an ID attribute
that must be a proper URI of the schema, and that can be used later
to extract a proper base URI for RefResolver.

This commit passes a path to OpenAPI spec written in URI notation as
spec ID, and uses RefResolver.from_schema(...) factory method to create
a proper resolver, which is enough to fix the original issue.

Fixes #3