Remove tox -e style env
There's no need to have a separate 'style' tox env to run flake8 linter
because we run it for every py{X} env before running tests with pytest.
That said, it's useless because if you run tests flake8 will be the
first check and will fail fast if some errors are found.