Add ':examples:' directive option
Thanks to awesome contributions, sphinxcontrib.openapi now supports
examples generation based on provided schema. However, the code turns
out to be buggy and is not well tested. That's why we don't want to
enable it by default, so we need to come up with an option that can
enables this path.