Add pre-commit hook
pre-commit became a de-facto standard for running various linters, so
let's start using it for yet-not-written code. All existing pieces are
effectively excluded in order to avoid merge conflicts with currently
open PRs. The plan is to remove these excludes separate one by one, as
the PRs are merged in.