Add renderers concept
The renderers in sphinxcontrib-openapi are the entities that render
OpenAPI spec. Each render may have its own set of options, and its own
rendering rules. So far we have only a single render but we believe this
is a way to go, as it allows to come up with better new implementations
without breaking changes.