Remove handling of optional import for mdinclude
Since commit 1e95ac6fc888 ("Actualize information in setup.py"), the
dependency on m2r, now replaced with mdinclude, is no longer part of the
"extras_require", but of the "install_requires" in setup.py. So the
dependency should always be satisfied, and we don't need to handle the
case where the import fails.